feat(deps): Use carat range for combined-stream
As mentioned in https://github.com/pelias/openaddresses/pull/343, we
were a bit nervous about a behavior change introduced in combined-stream 1.0.7.

However, as discovered in
https://github.com/felixge/node-combined-stream/issues/40 it's a bugfix
that makes previously unrecommended use of the library finally break.

The code in this repository is already using the library in the
recommended way, and isn't affected, so we might as well stay with the
latest version.

Closes https://github.com/pelias/csv-importer/pull/32