feat: supports for categories, multi-lang columns and aliases
# Background

The current CSV importer is very basic. We can do something much more interesting here, specially when we have custom data with many information.
OSM is a good example, when can have a tons of information like categories and multi-lang. This should be a good addition for this project.

# What's new ?

I added some new columns to enrich the data. Columns are:

- name_{lang}: Should be a ISO 639-1 language code, like we do with OSM
- name_json: Should be a JSON Array for the default name, this is cool for name aliases
- name_json_{lang}: Sould be a JSON Array for the specified lang
- category: A simple category to add for this entry
- category_json: A JSON Array to add many categories for an entry

# How it works ?

Use theses new columns in your CSV and voilà.